chore: rename primitives -> insights
