+++ packages/agent-context/src/insights/index.ts
